출력 및 처리함수 생성
