+++ README.md
@@ -1,6 +1,7 @@
 # 목표
-## 3/13
+## 3/13~14
 ### 실습 목표
+- [ ] 상태바 LightMode
 - [ ] View Background 이미지 설정
 - [ ] 카드 클래스 뒷면 이미지 설정
     - [ ] 크기 조절
@@ -18,3 +19,11 @@
     - [ ] [앱 표시 이름 변경](https://ldelight.tistory.com/m/189)
 - [ ] [Swift 프로세스 메모리 학습자료](https://www.youtube.com/watch?v=0Wnzb-0tjr4&feature=youtu.be)
 - [ ] Nested Enum 타입?
+
+# 결과
+## 1. 게임보드 만들기
+### 설명
+- Storyboard 없이 코드로 Assets에 저장한 이미지를 UIImageView에 넘겨줌
+- UIImageView의 Frame 하위 속성을 설정하여 원하는 위치에 배치
+### 출력 결과
+<img width="569" alt="Simulator Screen Shot - iPhone 14 Pro - 2023-03-15 at 15.25.00" src="https://user-images.githubusercontent.com/84387335/225236768-73c6f9fc-1e32-4dd4-acbd-a4edd0edad2b.png">
